// ABKC_API/Scripts/App/Models/Dog.d.ts
export interface DogModel {
    id: number;
    bullyId: number;
    abkcNo: string;
    ukcNo: string;
    akcNo?: any;
    adbaNo?: any;
    otherNo?: any;
    regNo: string;
    regType: string;
    dogName: string;
    title: string;
    suffix: string;
    breed: string;
    gender: string;
    color: string;
    sireNo: number;
    damNo: number;
    birthdate?: any;
    litterNo: number;
    ownerId: number;
    coOwnerId: number;
    tatooNo?: any;
    chipNo: string;
    points: number;
    champPoints: number;
    champWins: number;
    firstGeneration: boolean;
    registered: boolean;
    verified: boolean;
    printed: boolean;
    comments: string;
    dateCreated: string;
    dateRegistered?: any;
    regAmt: number;
    regUser: string;
    pups?: any;
    degrees?: any;
    lastModified: string;
    modifiedBy: string;
    bdna: string;
    nw: boolean;
    chcertPrinted: boolean;
    breedingDamOwnerId?: any;
    breedingDamCoOwnerId?: any;
    breedingSireOwnerId?: any;
    breedingSireCoOwnerId?: any;
    wpPoints: number;
    saveBully?: any;
    sabDate?: any;
    needsNewChcert: boolean;
    wpChampWins: number;
    needsNewWpchcert: boolean;
    wpTitle: string;
    n?: any;
    bb?: any;
    majors: number;
    judges: number;
    origTitle: string;
    inw?: any;
  }

  export interface BasicDogLookupInfo {
    id: int;
    dogName: string;
    bullyId: int;
    abkcNo: string;
  }